PROCESS: Acceptance tests refactored.
